move MemoryCondenser, LongTermMemory, json, out of the monologue
