+++ agenthub/monologue_agent/utils/prompts.py
@@ -5,6 +5,7 @@ from typing import List
 from opendevin.core import config
 from opendevin.core.exceptions import LLMOutputError
 from opendevin.core.schema.config import ConfigType
+from opendevin.core.utils import json
 from opendevin.events.action import (
     Action,
     action_from_dict,
@@ -12,8 +13,6 @@ from opendevin.events.action import (
 from opendevin.events.observation import (
     CmdOutputObservation,
 )
-
-from . import json
 
 ACTION_PROMPT = """
 You're a thoughtful robot. Your main task is this:
